Remove rebar erl_opts transform
This causes all dependencies to be compiled with the transform
which may not be available till the lager dependency is built.
This causes weird issues with clean build states and should be
avoided by putting the declaration inline.

At some point, this should probably be removed in favor of using
lager directly. The stats calls are probably better done with
specific backends in mind or related directly inline so that the
log calls themselves can be safely elided depending on the level
that is set.